clk: qcom: gpucc-msm8998: Properly acknowledge the required doubled clk speed for gpu_cc_pll0
With the flag CLK_SET_RATE_PARENT, the dependent clocks at which depend on gpu_cc_pll0 need to acknowledge the doubled rate required by the main parent clock. Thus, given this requirement we modify the frequency table for clk-rcg2 to double the rate of the GPU frequency it receives from devfreq. Because of this, we also need the FMAX to be doubled to match the parent clock rate.

In addition to this, allow gfx3d_clk to have access to GPU gpll0 clock and make the probe init functions static.

Signed-off-by: ederekun <sedrickvince@gmail.com>