clk: qcom: gpucc-msm8998: fix gfx3d_clk (again)
After removing the clk_ignore_unused command line flag, we were happy to
find out that the GPU stopped working.

Some time ago, we followed how 4.4 was managing this clock to get the
GPU working at full speed: now, after the latest changes to the
clocks, we need to partially revert the changes - in particular,
we need to get back to the 'not doubled/nominal' values for the VDD
on the gfx3d_clk clock.

Note that, even if the GPU seems to be working fine, two logs are
sometimes present:
- Rate 1030000195 for gfx3d_clk_src is greater than highest Fmax
- Rate 1420000195 for gfx3d_clk_src is greater than highest Fmax

It is not clear to me yet why these clocks are generated, as they
are over the maximum allowed speeds for gfx3d_clk (1030000000 or
1420000000).

Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>