fbdev: msm: improve MDSS_BROKEN_BOOTLOADER_HANDOFF
It turned out that, after the clks/regulators are reset due to the
MDSS_BROKEN_BOOTLOADER_HANDOFF config, the ESC clk is actually kept
active all the time.

This is causing the device to never enter RPM vmin mode, leading us to
higher power consumption.

For this reason, remove the vote on the ESC clk during dsi on.

To avoid calling this during device resume, perform this action only
once during boot.

Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>